[FIX] account: recompute line name when changing ref
Issue:
When changing the "Customer Reference" on an invoice, the name of the payment term line is not updated.

Steps to reproduce:
- Create an invoice with payment terms, confirm it
- Modify its "Customer Reference" to 'test' for example
- In the page "Journal Items" the name of the terms line has been recomputed to "test - INV/2025/XXXXX"
- Modify again its "Customer Reference" to 'abcdef' for example
- In the page "Journal Items" the name of the terms line was not recomputed

Cause:
In `_compute_name()` we only write the name if this condition is `True`:
```py
if n_terms > 1 or not line.name or line._origin.name == line._origin.move_id.payment_reference or (
        line._origin.move_id.payment_reference and line._origin.move_id.ref
        and line._origin.name == f'{line._origin.move_id.ref} - {line._origin.move_id.payment_reference}'
    ):
    line.name = name
```
The purpose of this line is to keep the name of the line if it was manually inputed. So the logic is:
we only write the computed name if the previous name was computed. To check this we check if
`line._origin.name == f'{line._origin.move_id.ref} - {line._origin.move_id.payment_reference}'`.

The issue comes from the use of `_origin` in a compute. `_origin` refers to the record before we
make any change. But it is meant to be used for `onchange` methods, in these the values are not yet written
so `_origin` refers to the record before saving.
Here, when saving, `line._origin` is the same as `line`, so
`line._origin.move_id.ref` is the new ref.
`line._origin.name` uses the old ref (it's currently being recomputed).

Solution:
Unfortunately, in the compute, there are no trace left of what were the previous values as the write
already occurred.

The initial complaint justifying to keep custom line names was that on bills, the line name is empty.
So when inputting a custom line name, it was removed by the compute method. The previous fix wanted to
be more general by always keeping custom line names.

Considering this, this commit removes part of the previous fix: Now we only keep the custom line name
when the compute method wants to remove it. So we keep the previous fix for bills.

There was a test verifying exactly that when manually deleting the line name, in the end
the line does not have a name. This will no longer be the case but a decision must be made between:
- updating the line name when changing the ref
- do not recompute the line name when it has been changed manually
- remove the line name when it's manually deleted
We can only keep two of them afaik

opw-5246917

closes odoo/odoo#239197

X-original-commit: 06417f7446a894e47f336aa0878018fcb710c50c
Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>
Signed-off-by: Mathieu Coutant (mcou) <mcou@odoo.com>