[FIX] mail: do not unpin sub-channel twice
Before this commit, outdated sub-channels were unpinned
each time the vacuum ran. It occurs because a condition
on sub-channel being pinned is missing.

In practice, it's not a big deal funtionnaly but leads
to useless notifications being sent.

While at it, this PR prevents unpinning sub-channels for
members who still have unread messages: the pin feature is
used to see unread messages on otherwise hidden threads.

closes odoo/odoo#239242

X-original-commit: 7fdc04f6d2ddb80fe2c50bfacb6cbb03b80647d0
Signed-off-by: Matthieu Stockbauer (tsm) <tsm@odoo.com>